Refactoring + Texture + XRBrush + SceneWindow 1. add XRCommon.h to include most commonly used headers (glm, stl, gl3w, etc) 2. add texture component support 3. implement XRBrush for primitive drawing 4. implement a draft version of scene window

+++ XRealityRender/src/simulation/SPHfluid/MarchCube.cpp
@@ -123,7 +123,7 @@ void MarchCube::computeGrid(const std::vector<SPHSim::SPHParticle> &particles, c
 	grid[thisGridIndex].flag = false; 
 
 
-	int dh = ceil(radius / grid_h);
+	int dh = int(ceil(radius / grid_h));
 
 	for (int dz = -dh; dz <= dh; dz++)
 	{
